requirements.txt
python-docx>=1.1.2
openpyxl>=3.1.5
pypdf>=5.9.0
